Trait linera_alloy_signer::SignerSync

pub trait SignerSync<Sig = Signature> {
    // Required methods
    fn sign_hash_sync(&self, hash: &B256) -> Result<Sig>;
    fn chain_id_sync(&self) -> Option<ChainId>;

    // Provided method
    fn sign_message_sync(&self, message: &[u8]) -> Result<Sig> { ... }
}

Synchronous Ethereum signer.

All provided implementations rely on sign_hash_sync. A signer may not always be able to implement this method, in which case it should return UnsupportedOperation, and implement all the signing methods directly.

Synchronous signers should also implement Signer, as they are always able to by delegating the asynchronous methods to the synchronous ones.

Required Methods

fn sign_hash_sync(&self, hash: &B256) -> Result<Sig>

Signs the given hash.

fn chain_id_sync(&self) -> Option<ChainId>

Returns the signer’s chain ID.

Provided Methods

fn sign_message_sync(&self, message: &[u8]) -> Result<Sig>

Signs the hash of the provided message after prefixing it, as specified in EIP-191.